group import of components in index.js refactor use of components in routes
