+++ src/components/index.js
@@ -1,1 +1,14 @@
-export { LoginUser } from './templates/login/LoginUser';
+import LoginUser from './templates/login/LoginUser';
+import LoginPassword from './templates/login/LoginPassword';
+import Transference1 from './templates/transferences/Transference1';
+import Transference2 from './templates/transferences/Transference2';
+import Transference3 from './templates/transferences/Transference3';
+
+export {
+  LoginUser,
+  LoginPassword,
+  Transference1,
+  Transference2,
+  Transference3,
+};
+
